refactor(#77): refactor dashboard to use react router dom
- add a router
- remove app.css and app.tsx and use main.tsx instead
- make a basic layout as a root
- add a trees router
